# Conflicts:
#	dist/DropZone.mpk
#	test/widgets/DropZone.mpk

+++ src/DropZone/widget/DropZone.js
@@ -5,7 +5,7 @@
     DropZone
     ========================
     @file      : Dropzone.js
-    @version   : 4.0.3
+    @version   : 4.0.5
     @author    : Andries Smit & Chris de Gelder
     @date      : 06-09-2017 
     @license   : Apache V2
@@ -16,6 +16,7 @@
 	Mendix 7.12 version.
 	- 8-3-2018 mxui.dom. functions replaced with domConstruct.create
 	- 8-3-2018 Merge csrftoken (Thanks Jelte)
+	- 11-5-2018 disabled logger messages, bug in 14.1
     
     To be done:
     - fix, upload button image
@@ -56,7 +57,6 @@ define([
          */
         constructor: function () {
             logger.debug(this.id + ".constructor");
-			logger.debug('dropzone', Constructdropzone);
             this.dropzone = null;
             this._contextObj = null;
         },
@@ -67,7 +67,6 @@ define([
         postCreate: function () {
             logger.debug(this.id + ".postCreate");
             this.initDropZone();
-			logger.debug('this', this);
         },
         /**
          * mxui.widget._WidgetBase.update is called when context is changed or initialized. Implement to re-render and / or fetch data.
@@ -78,7 +77,9 @@ define([
         update: function (obj, callback) {
             logger.debug(this.id + ".update");
             this._contextObj = obj;
-            mendix.lang.nullExec(callback);
+            if (callback) {
+				callback();
+			}
         },
         /**
          * initalize the dropzone library. 
@@ -129,7 +130,7 @@ define([
             this.dropzone.on("error", dojoLang.hitch(this, this.onError));
             this.dropzone.on("removedfile", dojoLang.hitch(this, this.onRemoveFile));
 			this.dropzone.on("sending", dojoLang.hitch(this, this.addFormData));
-			
+            logger.debug(this.id + ".initDropZone done");
         },
         /**
          * add Mendix 7 'data' part to formdata
@@ -141,7 +142,8 @@ define([
 		addFormData: function(data, xhr, formData) {
 			// Mendix 7 expects a data part.
 			var s = '{"changes":{},"objects":[]}';
-			formData.append("data", s);			
+			formData.append("data", s);
+			//formData.append("data", JSON.stringify( { changes: {}, objects: [] }));			
 		},
         /**
          * set the Mendix upload URL based on the GUID
@@ -160,7 +162,7 @@ define([
          */
         onError: function (file, message) {
             logger.error(this.id + ".onError", message);
-            this.onRemoveFile(file);
+            this.removeFile(file);
         },
         /**
          * an image should be removed from within a microflow, if there is non just delete if via the api
@@ -168,31 +170,34 @@ define([
          * @param {type} message - status message
          * @returns {undefined}
          */
-        onRemoveFile: function (file, message) {
+		onRemoveFile: function (file, message) {
             if (this._beingDestroyed) {
                 // dont remove the files when the widget is being destroyed by the uninitialize function.
                 return;
             }
             logger.debug(this.id + ".onRemoveFile");
             var obj = file.obj;
-            if (obj && this.onRemove) {
-                mx.data.action({
-                    params: {
-                        actionname: this.onRemove,
-                        applyto: "selection",
-                        guids: [obj.getGuid()]
-                    },
-                    origin: this.mxform,
-                    callback: dojoLang.hitch(this, function (result) {
-                        file.obj = null;
-                    }),
-                    error: function (e) {
-                        logger.error("onRemoveFile", e);
-                    }
-                });
-            } else {
-                this.removeFile(file);
-            }
+			// if autoremoveafter upload is enabled the removefile is called but should not remove the file from the Mendix application
+			if (!file.deleteAfterUpload) {
+				if (obj && this.onRemove) {
+					mx.data.action({
+						params: {
+							actionname: this.onRemove,
+							applyto: "selection",
+							guids: [obj.getGuid()]
+						},
+						origin: this.mxform,
+						callback: dojoLang.hitch(this, function (result) {
+							file.obj = null;
+						}),
+						error: function (e) {
+							logger.error("onRemoveFile", e);
+						}
+					});
+				} else {
+					this.removeFile(file);
+				}
+			}
         },
         /**
          * when uploadload is completed, commit and call onchange MF
@@ -208,6 +213,10 @@ define([
                     callback: dojoLang.hitch(this, function () {
                         logger.debug("onComplete");
                         this.callOnChange(file.obj);
+						if (this.removeAfterUpload) {
+							file.deleteAfterUpload = true;
+							this.dropzone.removeFile(file);
+						}
                     })
                 });
             }
@@ -215,6 +224,7 @@ define([
 				this.dropzone.processQueue(); 
 			}
         },
+		
         /**
          * Create file on mendix server, and validate if it could be accepted.
          * @param {File} file - the file that validate
@@ -273,13 +283,10 @@ define([
                 callback: dojoLang.hitch(this, function (obj) {
 					logger.debug('create', obj);
                     var ref = this.contextassociation.split("/");
-                    if (obj.has(ref[0])) {
+                    if (obj.has(ref[0]) && this._contextObj) {
                         obj.set(ref[0], this._contextObj.getGuid());
                     }
                     obj.set(this.nameattr, file.name);
-                    if (this.sizeattr) {
-                        obj.set(this.sizeattr, file.size);
-                    }
                     if (this.typeattr) {
                         obj.set(this.typeattr, file.type);
                     }
@@ -336,7 +343,6 @@ define([
             logger.debug(this.id + ".onclickEvent");
             this.dropzone.processQueue(); 
 			logger.debug('dz', this.dropzone.getQueuedFiles()); 
-			//cdg test this.processQueue();
         },
         /**
          * Call onchange Miroflow if any.
